1 MB limit to be removed and text to be adjusted
Item Variation - Description
Quantity Sold
Item Variation - New Additions to DB - All should have validations
CK Editor things.
Inventory Item - Average Cost
Total Quantity in Hand = Sum (Qty from Manufacturers) + Qty in Hand
Quantity in Package on Inventory Item comes from Item Variation
Inventory Image - 10 MB Limit

+++ EPMS.Web/Models/InventoryItem.cs
@@ -34,7 +34,7 @@ namespace EPMS.Web.Models
         public double? AverageCost { get; set; }
         public double? AveragePackagePrice { get; set; }
         public double? QuantityInPackage { get; set; }
-        public long? QuantityInHand { get; set; }
+        public double? QuantityInHand { get; set; }
         public long? QuantitySold { get; set; }
         public string DepartmentPath { get; set; }
